requirements.txt
boto3
matplotlib
python-dotenv